feat: demo 5 start

+++ 01fundamentals/05StringsStreamsDemos/03StdStringBasics.cpp
@@ -1,18 +1,21 @@
 #include <iostream>
-#include <string>
+// #include <string>
 using namespace std;
 
 int main() {
   string theFoxPart = "the quick brown fox";
   string theActionPart("jumps over");
+
   char dogPartCString[] = "the lazy dog";
+
   string sentence = theFoxPart + string("---") + theActionPart +
                     string(3, '-') + dogPartCString;
+
   string sameSentance = "the quick brown fox---jumps over---the lazy dog";
   cout << sentence << " (length: " << sentence.size() << ")" << endl;
 
   string hello = "hello";
-  for (int i = 0; i < hello.size(); i++) {
+  for (size_t i = 0; i < hello.size(); i++) {
     cout << hello[i] << endl;
   }
